pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-rest-votings"
version = "0.0.1"
readme = "README.md"
license = {file = "LICENSE"}
description = ""
requires-python = ">=3.8"
dependencies = [
  "Django >= 4.2",
  "djangorestframework",
]
authors = [
  { name = "My Name", email = "me@email.com" },
]
keywords = ["example1", "example2"]

[project.urls]
Homepage = "https://github.com/carlos-m-a/django-rest-votings"
"Bug Tracker" = "https://github.com/carlos-m-a/django-rest-votings/issues"


[tool.setuptools.packages.find]
include = ["votings*"]